提交本地

+++ hyjf_am_server/hyjf_am_core/hyjf_am_trade/src/main/java/com/hyjf/am/trade/dao/mapper/customize/HjhPlanCustomizeMapper.java
@@ -132,7 +132,7 @@ public interface HjhPlanCustomizeMapper {
      * @param params
      * @return
      */
-    Long selectPlanAccedeSum(Map<String,Object> params);
+    Integer selectPlanAccedeSum(Map<String,Object> params);
 
     /**
      * 查询相应的计划的加入明细
