+++ hyjf_cs_server/hyjf_cs_trade/src/main/java/com/hyjf/cs/trade/client/AmTradeClient.java
@@ -2137,7 +2137,7 @@ public interface AmTradeClient {
 
 	String getborrowIdByProductId(Map<String, Object> params);
 
-	Long countBatchCenter(BatchCenterCustomize batchCenterCustomize);
+    Integer countBatchCenter(BatchCenterCustomize batchCenterCustomize);
 
 	List<BatchCenterCustomizeVO> selectBatchCenterList(BatchCenterCustomize batchCenterCustomize);
