+++ hyjf_cs_server/hyjf_cs_trade/src/main/java/com/hyjf/cs/trade/service/repay/impl/BatchHjhBorrowRepayApiServiceImpl.java
@@ -22,7 +22,7 @@ public class BatchHjhBorrowRepayApiServiceImpl extends BaseServiceImpl implement
 	 * countBatchCenter
 	 */
 	@Override
-	public Long countBatchCenter(BatchCenterCustomize batchCenterCustomize) {
+	public Integer countBatchCenter(BatchCenterCustomize batchCenterCustomize) {
 		return borrowClient.countBatchCenter(batchCenterCustomize);
 	}
